requirements.txt
discord.py
python-dotenv
requests
asyncpg
asyncio
pillow
googletrans
firebase_admin
tatsumaki.py